+++ gulpfile.js
@@ -1,1 +1,8 @@
-require('./src').initialize(require('gulp'));
+'use strict';
+
+const gulp = require('gulp');
+const tools = require('./src');
+
+tools.initialize(gulp);
+
+gulp.task('default', ['pre-release']);