+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 - Introduce globals (common configurations used by multiple tasks)
+- Extended README (usage section)
+- Initialize dummy default task
 
 ## 0.1.8 (2016-01-15)
 - Allow tasks to be disabled
@@ -11,7 +13,6 @@
 - Enhance configuration support of the webpack task
 - Add legacy decorator support to babel
 - Expose default configuration
-- Dummy default task
 
 ## 0.1.6 (2016-01-08)
 - Proper JSDoc documentation